requirements.txt
flask
requests
google-generativeai
gunicorn
werkzeug